+++ verifier-backend/src/test/java/ch/post/it/evoting/verifier/backend/verifications/setup/evidence/VerifyVotingOptionsAlgorithmTest.java
@@ -38,9 +38,11 @@ import ch.post.it.evoting.cryptoprimitives.math.GroupVector;
 import ch.post.it.evoting.cryptoprimitives.math.PrimeGqElement;
 import ch.post.it.evoting.evotinglibraries.domain.VotingOptionsConstants;
 
+@SuppressWarnings("java:S116")
 @DisplayName("VerifyVotingOptionsAlgorithm calling verifyVotingOptions with")
 class VerifyVotingOptionsAlgorithmTest {
 
+	@SuppressWarnings("java:S116") // Aligns with our coding conventions for algorithms.
 	private final int n_sup = VotingOptionsConstants.MAXIMUM_SUPPORTED_NUMBER_OF_VOTING_OPTIONS;
 	private final VerifyVotingOptionsAlgorithm verifyVotingOptionsAlgorithm = new VerifyVotingOptionsAlgorithm();