+++ verifier-backend/src/test/java/ch/post/it/evoting/verifier/backend/verifications/tally/authenticity/VerifySignatureTallyComponentVotesTest.java
@@ -123,6 +123,7 @@ class VerifySignatureTallyComponentVotesTest extends TallyVerificationTest {
 		assertEquals(expected, getHashTallyComponentVotesSpec(tallyComponentVotesPayload));
 	}
 
+	@SuppressWarnings("java:S117")
 	private String getHashTallyComponentVotesSpec(final TallyComponentVotesPayload tallyComponentVotesPayload) {
 		final HashableString ee = HashableString.from(tallyComponentVotesPayload.getElectionEventId());