+++ verifier-frontend/src/app/shared/components/verification-details/verification-details.component.ts
@@ -17,9 +17,9 @@ import {Component, Input} from '@angular/core';
 import {KeyValuePipe, NgClass, SlicePipe} from "@angular/common";
 import {LocalizePipe} from "../../pipes/localize.pipe";
 import {VerificationDefinition} from "../../types/verification-definition";
-import {VerificationStatus} from "../../types/verification-status";
 import {TranslatePipe} from "@ngx-translate/core";
 import {NgbCollapseModule} from "@ng-bootstrap/ng-bootstrap";
+import {errorStack, isDefault, isError, isNotOK, isOK, isRunning, showErrorStack} from "../../utils/verification-utils";
 
 @Component({
   templateUrl: 'verification-details.component.html',
@@ -35,42 +35,20 @@ import {NgbCollapseModule} from "@ng-bootstrap/ng-bootstrap";
   standalone: true
 })
 export class VerificationDetailsComponent {
-
   @Input() verifications: Record<string, VerificationDefinition>;
   @Input() isReportDisplay: boolean;
 
   collapsedStates: boolean[] = [];
 
+  protected readonly isError = isError;
+  protected readonly isNotOK = isNotOK;
+  protected readonly showErrorStack = showErrorStack;
+  protected readonly isOK = isOK;
+  protected readonly isRunning = isRunning;
+  protected readonly isDefault = isDefault;
+  protected readonly errorStack = errorStack;
+
   toggleCollapse(index: number) {
     this.collapsedStates[index] = !this.collapsedStates[index];
   }
-
-  isDefault(key: unknown): boolean {
-    return this.verifications[<string> key].status === VerificationStatus.IDLE;
-  }
-
-  isRunning(key: unknown): boolean {
-    return this.verifications[<string> key].status === VerificationStatus.RUNNING;
-  }
-
-  isOK(key: unknown): boolean {
-    return this.verifications[<string> key].status === VerificationStatus.OK;
-  }
-
-  isNotOK(key: unknown): boolean {
-    return this.verifications[<string> key].status === VerificationStatus.NOK;
-  }
-
-  isError(key: unknown): boolean {
-    return this.verifications[<string> key].status === VerificationStatus.ERROR;
-  }
-
-  showErrorStack(key: unknown): boolean {
-    return this.verifications[<string> key].errorStack !== null && !this.isReportDisplay;
-  }
-
-  errorStack(key: unknown): string[] {
-    return this.verifications[<string> key].errorStack || [];
-  }
-
 }